Rename pager to obsolete pager
